+++ components/01-globals/icons/icons.config.js
@@ -19,11 +19,27 @@ module.exports = {
         class: 'icons__single-icon'
       },
       {
-        iconId: 'arrow-circle-up',
+        iconId: 'angle-right',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'angle-left',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'arrow-down',
         class: 'icons__single-icon'
       },
       {
         iconId: 'arrow-left',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'arrow-right',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'arrow-up',
         class: 'icons__single-icon'
       },
       {
@@ -32,6 +48,10 @@ module.exports = {
       },
       {
         iconId: 'close',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'compare',
         class: 'icons__single-icon'
       },
       {
@@ -55,6 +75,10 @@ module.exports = {
         class: 'icons__single-icon'
       },
       {
+        iconId: 'grid',
+        class: 'icons__single-icon'
+      },
+      {
         iconId: 'heart',
         class: 'icons__single-icon'
       },
@@ -64,6 +88,10 @@ module.exports = {
       },
       {
         iconId: 'linkedin',
+        class: 'icons__single-icon'
+      },
+      {
+        iconId: 'list',
         class: 'icons__single-icon'
       },
       {